create relations to get results of votes with answer and user ids, disabled groups to prevent issues

+++ app/Http/Controllers/RoleController.php
@@ -132,6 +132,8 @@ class RoleController extends Controller
 
         $this->authorize('delete', $role);
 
+        if ($role->users()->first()) return redirect()->route('roles.index')->with('roleDeleteFailure', "Ce rôle ne peut pas être supprimé");
+
         $role->permissions()->detach($role->permissions()->pluck('id')->toArray());
         $role->delete();
